Git-tag command generation, spec alteration, generations split to command.rb
