+++ app/spec/lib/bk/compat/bitrise/examples/git-tag.yml
@@ -17,8 +17,25 @@ workflows:
         - tag: $BITRISE_BUILD_NUMBER
         - tag_message: "v1.0.0"
         - push: true
+  test:
+    steps:
+    - git-clone@8.2.2:
+        inputs:
+        - clone_into_dir: /tmp/bitrise-ex/
+        - repository_url: git@github.com:example/example-repository.git
+    - script@1.1.5:
+        inputs:
+        - content: ./test.sh
+    - git-tag@1.1.0:
+        inputs:
+        - tag: test-$BITRISE_BUILD_NUMBER
+        - push: true
   deploy:
     steps:  
     - script@1.1.5:
         inputs:
         - content: ./deploy-app.sh
+    - git-tag@1.1.0:
+        inputs:
+        - tag: deploy-$BITRISE_BUILD_NUMBER
+        - tag_message: "v1.0.0"
